Status bar: Battery customizations, (1/2):
- Show/hide battery
- Show/hide battery text
- Enable/disable cut out battery text
- Custom battery color
- Custom text color

This also takes effect of the status bar when on the lock screen,
but no effect on the battery of the expanded header.
The SystenUI Tuner setting ("Show embedded battery percentage")
takes only effect on the battery of the expanded header.

Additonal changes:
The battery level text of the status bar when on the lock screen
will only be shown, if the battery text is hidden, battery is shown and
the battery is charging.

The dark mode code is already included, but unfortunatelly, I noticed
it isn`t working as expected, (also for the clock, network traffic and carrier label),
so it`s disabled for now.
I`ll disable the dark mode for clock, network traffic and carrier label with the next commit, too.
I guess the dark mode needs some more overwork to work as expected.

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/StatusBarIconController.java

Conflicts:
	core/java/android/provider/Settings.java