Core: Use ro.build.date to signal mIsUpgrade
Modified the original commit by dhacker29 to check for build date

Original commit by dhacker29:
Core: Use ro.purenexus.version to signal mIsUpgrade
We use a static fingerprint that is only changed when a new stock
version is released, so every flash shows Android is starting instead
of upgrading.  This will fix that.

Change-Id: Ic5559b662f5e6fc4df1f6929494663bcb84f8333